pyproject.toml
[project]
name = "python-playwright"
version = "0.1.0"
description = "Playwright fun"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pylint>=3.3.7",
    "pytest==8.4.1",
    "pytest-playwright==0.7.0",
    "testcontainers==4.10.0",
]
